Added new ValidityAwareTaghelperBase to encapsulate validation-related aspects of the base tag helper
